Move ExpandableListAdapter
